python_api/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
mysql-connector-python>=8.2.0
bcrypt>=4.1.0
